Development Merge 05/03/25 - Initial Setup (#1)
Setup the baseline UE5 project, added the top-down template as a kick-off point, edited the gitignore as well.

* [TASK #1 & #2] Initial Project Setup

Setup the Github Repository for Git Flow.
Created the initial UE5 project and imported the top-down template, to be modified with the correct controls.
Setup the gitignore to reduce size of repository and fetched files.

* [TASK #1 & #2] Initial Project Setup

Setup the Github Repository for Git Flow.
Created the initial UE5 project and imported the top-down template, to be modified with the correct controls.
Setup the gitignore to reduce size of repository and fetched files.