EQS added
The enemy ai now makes use of EQS for better positioning around the player when attacking. After attacking it backs away and strafes around the player. Also made a replacement for the standard move to task that includes idea attacking and defending ranges. Observer aborts need a looking at as sometimes it locks onto the strafe behaviour when the player moves  away.