Picking Weapons UI & Folder Sorting
Added the weapon picking screen, this is currently setup through the BPC itself so it will ALWAYS activate through it. Should be fine as is currently but might cause some issues later down the line

Made it so that the (weapon character) player can't try to auto-aim instantly upon spawning in by disabling input & Enabling once the weapon has been picked.

Did some testing and the weapon can be changed on the fly if needed like Hade's selection room, just will need to be changed a bit through the Ui but should be fine.

Sorted out some of my folders so they aren't confusing as much anymore, additionally added some comments to some blueprints that didn't have any previously.