Updated Aiming, added charge on hold functions in.
Updated the aiming with help from Emily!

Added a charge on hold function which rotates towards the mouse so you can aim when holding down stuff, additionally added a "perfect" shield bash on the sword, so you can perfect the amount and deal more damage, in addition to this i've added two more boxes on the sword for blocking, one for a normal shield bash and one for the perfect one. These activate upon letting go of the block holding button.