Weapon Meshes, UI Fixes, UI Icons, decals in progress
I've updated the weapon meshes and all of them are ingame and working properly (minus the variants messing up anims still)

Also added icons to the variants UI showcasing their unqiue icons.

And have developed the decal more, it slightly works but it needs a start and end point to go to instead of just expanding forward.